factorise data_merger
